Renamed read.py to read_apis.py for consistency, and modified all former references to read.py accordingly.
